big refactoring 3
