+++ src/cart/cart.module.ts
@@ -7,7 +7,7 @@ import { CartService } from './services';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Carts } from './entity/Carts';
 import { CartItems } from './entity/CartItems';
-import { Orders } from '../order/entryty/Order';
+import { Orders } from '../order/entity/Order';
 
 @Module({
   imports: [OrderModule, TypeOrmModule.forFeature([Carts, CartItems, Orders])],
