+++ src/order/order.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { OrderService } from './services';
 import { OrderController } from './order.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Orders } from './entryty/Order';
+import { Orders } from './entity/Order';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Orders])],
