+++ src/pg-scripts/tables.sql
@@ -2,12 +2,20 @@ create type statusType as enum ('OPEN', 'ORDERED')
 
 create type paymentType
 
+-- create table carts (
+-- 	id uuid primary key default uuid_generate_v4(),
+-- 	user_id uuid,
+-- 	created_at date not null,
+-- 	updated_at date not null,
+-- 	status statusType
+-- )
+
 create table carts (
 	id uuid primary key default uuid_generate_v4(),
 	user_id uuid,
-	created_at date not null,
-	updated_at date not null,
-	status statusType
+	created_at date,
+	updated_at date,
+	status text
 )
 
 -- create table cart_items (
@@ -22,7 +30,7 @@ create table cart_items (
     cart_id uuid,
     product_id uuid,
     count integer,
-    price integer,
+    price NUMERIC(9,2),
     foreign key ("cart_id") references "carts" ("id")
 )
 
@@ -33,7 +41,7 @@ create table orders (
     payment json,
     delivery json,
     comments text,
-    status statusType,
+    status text,
     total integer,
     foreign key ("cart_id") references "carts" ("id")
 )
